Remove more duplicated code, create utils.map
