+++ erftools/data/US_state_borders_coordinates.txt
